+++ CHANGES.md
@@ -6,6 +6,8 @@ Change Log
 * Pin `html-to-react` to `1.4.5` due to ESM module in dependency (`parse5`) breaking webpack
 * Add step to `"Deploy TerriaMap"` action to save `yarn.lock` after `sync-dependencies` (for debug purposes)
 * Implement ability to navigate between scenes in StoryPanel using keyboard arrows
+* Fix "Regions: xxx" short report showing for non region mapped items
+* Fix `showInChartPanel` default for mappable items
 * [The next improvement]
 
 #### 8.2.8 - 2022-07-04